+++ include/swift/AST/TypeAlignments.h
@@ -35,9 +35,11 @@ namespace swift {
   class CaptureListExpr;
   class Decl;
   class DeclContext;
+  class DifferentiableAttr;
   class Expr;
   class ExtensionDecl;
   class GenericEnvironment;
+  class GenericParamList;
   class GenericTypeParamDecl;
   class NominalTypeDecl;
   class NormalProtocolConformance;
@@ -46,9 +48,12 @@ namespace swift {
   class Pattern;
   class ProtocolDecl;
   class ProtocolConformance;
+  class RequirementRepr;
   class SILFunction;
   class SILFunctionType;
+  class SpecializeAttr;
   class Stmt;
+  class TrailingWhereClause;
   class TypeVariableType;
   class TypeBase;
   class TypeDecl;
@@ -63,9 +68,15 @@ namespace swift {
   constexpr size_t StmtAlignInBits = 3;
   constexpr size_t TypeAlignInBits = 3;
   constexpr size_t PatternAlignInBits = 3;
-  constexpr size_t SILFunctionAlignInBits = 2;
-  constexpr size_t TypeVariableAlignInBits = 4;
   constexpr size_t TypeReprAlignInBits = 3;
+
+  constexpr size_t SILFunctionAlignInBits = 2;
+  constexpr size_t ASTContextAlignInBits = 2;
+  constexpr size_t TypeVariableAlignInBits = 4;
+
+  // Well, this is the *minimum* pointer alignment; it's going to be 3 on
+  // 64-bit targets, but that doesn't matter.
+  constexpr size_t PointerAlignInBits = 2;
 }
 
 namespace llvm {
@@ -110,8 +121,9 @@ LLVM_DECLARE_TYPE_ALIGNMENT(swift::SILFunctionType,
 LLVM_DECLARE_TYPE_ALIGNMENT(swift::Stmt, swift::StmtAlignInBits)
 LLVM_DECLARE_TYPE_ALIGNMENT(swift::BraceStmt, swift::StmtAlignInBits)
 
-LLVM_DECLARE_TYPE_ALIGNMENT(swift::ASTContext, 2);
+LLVM_DECLARE_TYPE_ALIGNMENT(swift::ASTContext, swift::ASTContextAlignInBits);
 LLVM_DECLARE_TYPE_ALIGNMENT(swift::DeclContext, swift::DeclContextAlignInBits)
+LLVM_DECLARE_TYPE_ALIGNMENT(swift::DifferentiableAttr, swift::PointerAlignInBits)
 LLVM_DECLARE_TYPE_ALIGNMENT(swift::Expr, swift::ExprAlignInBits)
 LLVM_DECLARE_TYPE_ALIGNMENT(swift::CaptureListExpr, swift::ExprAlignInBits)
 LLVM_DECLARE_TYPE_ALIGNMENT(swift::AbstractClosureExpr, swift::ExprAlignInBits)
@@ -121,10 +133,17 @@ LLVM_DECLARE_TYPE_ALIGNMENT(swift::NormalProtocolConformance,
                             swift::DeclAlignInBits)
 LLVM_DECLARE_TYPE_ALIGNMENT(swift::GenericEnvironment,
                             swift::DeclAlignInBits)
+LLVM_DECLARE_TYPE_ALIGNMENT(swift::GenericParamList,
+                            swift::PointerAlignInBits)
 LLVM_DECLARE_TYPE_ALIGNMENT(swift::Pattern,
                             swift::PatternAlignInBits)
+LLVM_DECLARE_TYPE_ALIGNMENT(swift::RequirementRepr,
+                            swift::PointerAlignInBits)
 LLVM_DECLARE_TYPE_ALIGNMENT(swift::SILFunction,
                             swift::SILFunctionAlignInBits)
+LLVM_DECLARE_TYPE_ALIGNMENT(swift::SpecializeAttr, swift::PointerAlignInBits)
+LLVM_DECLARE_TYPE_ALIGNMENT(swift::TrailingWhereClause,
+                            swift::PointerAlignInBits)
 
 LLVM_DECLARE_TYPE_ALIGNMENT(swift::AttributeBase, swift::AttrAlignInBits)
