+++ lib/IDE/Formatting.cpp
@@ -11,6 +11,7 @@
 //===----------------------------------------------------------------------===//
 
 #include "swift/AST/ASTWalker.h"
+#include "swift/AST/GenericParamList.h"
 #include "swift/AST/TypeRepr.h"
 #include "swift/IDE/SourceEntityWalker.h"
 #include "swift/Parse/Parser.h"
